Cleaning up some whitespace

+++ Controllers/TeamController.cs
@@ -2,27 +2,23 @@
 using IssueTracker.ViewModels;
 using System;
 using System.Collections.Generic;
-using System.Dynamic;
 using System.Web.Mvc;
-using System.Web;
 
 namespace IssueTracker.Controllers
 {
     public class TeamController : Controller
     {
-        private Repository.TeamRepository TeamRepository = new Repository.TeamRepository();
-        private Repository.TeamViewRepository TeamViewRepository = new Repository.TeamViewRepository();
-        private Repository.UserRepository UserRepository = new Repository.UserRepository();
-        private Repository.UserTeamRoleRepository UserTeamRoleRepository = new Repository.UserTeamRoleRepository();
-        private Repository.TeamGroupRepository TeamGroupRepository = new Repository.TeamGroupRepository();
-
-        // GET: Team
+        private readonly Repository.TeamRepository TeamRepository = new Repository.TeamRepository();
+        private readonly Repository.TeamViewRepository TeamViewRepository = new Repository.TeamViewRepository();
+        private readonly Repository.UserRepository UserRepository = new Repository.UserRepository();
+        private readonly Repository.UserTeamRoleRepository UserTeamRoleRepository = new Repository.UserTeamRoleRepository();
+        private readonly Repository.TeamGroupRepository TeamGroupRepository = new Repository.TeamGroupRepository();
         public ActionResult Index()
         {
             try
             {
                 List<TeamModel> userTeamRoleModels = new List<TeamModel>();
-                userTeamRoleModels = TeamRepository.GetTeamsCreatedBy();
+                userTeamRoleModels = TeamRepository.GetTeamsForCurrentUserId(UserRepository.GetCurrentUser().UserId);
                 return View("Index", userTeamRoleModels);
             }
             catch
@@ -30,11 +26,12 @@ namespace IssueTracker.Controllers
                 return View("Index","Home");
             }
         }
-
         public ActionResult EditTeamMember(Guid TeamId, Guid UserId)
         {
             TeamViewModel teamViewModel = new TeamViewModel();
             teamViewModel = TeamViewRepository.GetTEamViewModelsByTeamId(TeamId).Find(x => x.UserId == UserId);
+            ViewBag.CreatedBy = TeamRepository.GetTeamById(TeamId).CreatedBy;
+            ViewBag.CurrentUser = UserRepository.GetCurrentUser().UserId;
             return View("EditTeamMember",teamViewModel);
         }
         [HttpPost]
@@ -46,7 +43,6 @@ namespace IssueTracker.Controllers
             TeamViewRepository.UpdateTeamGroup(teamViewModel);
             return RedirectToAction("Details", new { id=teamViewModel.TeamId});
         }
-
         public ActionResult AddUser(Guid id)
         {
             TeamViewModel teamViewModel = new TeamViewModel();
@@ -61,14 +57,11 @@ namespace IssueTracker.Controllers
                 return View("AddUser", teamViewModel);
             }
         }
-
         [HttpPost]
         public ActionResult AddUser(FormCollection collection)
         {
-            dynamic teamGroups = new ExpandoObject();
             try
-            {      //check if users exists in team and change only ream role if necesarry
-
+            {
                 TeamViewModel teamViewModel = new TeamViewModel();
                 UpdateModel(teamViewModel);
                 TeamViewRepository.CreateTeamGroup(teamViewModel);
@@ -78,40 +71,31 @@ namespace IssueTracker.Controllers
             {
                 return View();
             }
-
-
         }
-        
-        // GET: Team/Details/5
-        public ActionResult Details(Guid id)
+        public ActionResult Details(Guid TeamId)
         {
-            List<TeamViewModel> teamViewModels = TeamViewRepository.GetTEamViewModelsByTeamId(id);
-            if (TeamViewRepository.GetTEamViewModelsByTeamId(id).Count == 0)
+            List<TeamViewModel> teamViewModels = TeamViewRepository.GetTEamViewModelsByTeamId(TeamId);
+            ViewBag.CreatedBy = TeamRepository.GetTeamById(TeamId).CreatedBy;
+            ViewBag.CurrentUser = UserRepository.GetCurrentUser().UserId;
+            if (TeamViewRepository.GetTEamViewModelsByTeamId(TeamId).Count == 0)
             {
-                TeamViewModel teamViewModel = new TeamViewModel();
-                teamViewModel.TeamId = id;
-                teamViewModel.UserNameSurname = "";
+                TeamViewModel teamViewModel = new TeamViewModel
+                {
+                    TeamId = TeamId,
+                    UserNameSurname = ""
+                };
                 teamViewModels.Add(teamViewModel);
                 return View("Details", teamViewModels);
             }
-            
             return View("Details", teamViewModels);
         }
-
-        // GET: Team/Create
         public ActionResult Create()
         {
             return View();
         }
-
-
-
-
-        // POST: Team/Create
         [HttpPost]
         public ActionResult Create(FormCollection collection)
         {
-            dynamic teamGroups = new ExpandoObject();
             try
             {
                 TeamModel teamModel = new TeamModel();
@@ -119,21 +103,16 @@ namespace IssueTracker.Controllers
                 TeamRepository.CreateTeam(teamModel);
                 return RedirectToAction("Index");
             }
-            catch
+            catch(Exception ex)
             {
-                return View();
+                return View(ex);
             }
         }
-
-        // GET: Team/Edit/5
         public ActionResult EditTeam(Guid TeamId)
         {
-            TeamModel teamModel = new TeamModel();
-            teamModel = TeamRepository.GetTeamById(TeamId);
+            TeamModel teamModel = TeamRepository.GetTeamById(TeamId);
             return View("Edit",teamModel);
         }
-
-        // POST: Team/Edit/5
         [HttpPost]
         public ActionResult EditTeam(Guid TeamId, FormCollection collection)
         {
@@ -149,23 +128,38 @@ namespace IssueTracker.Controllers
                 return View();
             }
         }
-
-        // GET: Team/Delete/5
         public ActionResult DeleteTeam(Guid TeamId)
         {
-            ViewBag.TeamName = TeamRepository.GetTeamsCreatedBy().Find(x => x.TeamId == TeamId).TeamName;
+            if (TeamRepository.GetTeamsCreatedBy().Find(x => x.TeamId == TeamId) != null)
+            {
+                if (UserRepository.GetCurrentUser().UserId == TeamRepository.GetTeamsCreatedBy().Find(x => x.TeamId == TeamId).CreatedBy)
+                {
+                    ViewBag.TeamName = TeamRepository.GetTeamsCreatedBy().Find(x => x.TeamId == TeamId).TeamName;
+                    return View();
+                }
+                return View();
+            }
             return View();
         }
-
         [HttpPost]
         public ActionResult DeleteTeam(Guid TeamId, FormCollection collection)
         {
             try
             {
-                TeamModel teamModel = new TeamModel();
-                teamModel.TeamId = TeamId;
-                TeamRepository.DeleteTeam(teamModel);
+                if (TeamRepository.GetTeamsCreatedBy().Find(x => x.TeamId == TeamId) != null)
+                {
+                    if (UserRepository.GetCurrentUser().UserId == TeamRepository.GetTeamsCreatedBy().Find(x => x.TeamId == TeamId).CreatedBy)
+                    {
+                        TeamModel teamModel = new TeamModel
+                        {
+                            TeamId = TeamId
+                        };
+                        TeamRepository.DeleteTeam(teamModel);
 
+                        return RedirectToAction("Index");
+                    }
+                    return RedirectToAction("Index");
+                }
                 return RedirectToAction("Index");
             }
             catch
@@ -173,24 +167,23 @@ namespace IssueTracker.Controllers
                 return View();
             }
         }
-        // GET: Team/Details/AddUser/Delete/5
         public ActionResult DeleteUserFromTeam(Guid TeamId, Guid UserId, Guid TeamRoleId)
         {
             ViewBag.UserName = UserRepository.GetUserById(UserId).UserName;
-            ViewBag.TeamRoleName = UserTeamRoleRepository.GetTeamRoleModels().Find(x => x.userTeamRoleId == TeamRoleId).UserTeamRoleName;
+            ViewBag.TeamRoleName = UserTeamRoleRepository.GetTeamRoleModels().Find(x => x.UserTeamRoleId == TeamRoleId).UserTeamRoleName;
             return View();
         }
-
-        // POST: Team/Details/AddUser/Delete/5
         [HttpPost]
-        public ActionResult DeleteUserFromTeam(Guid TeamId,Guid UserId,Guid TeamRoleId, FormCollection collection)
+        public ActionResult DeleteUserFromTeam(Guid TeamId, Guid UserId, Guid TeamRoleId, FormCollection collection)
         {
             try
             {
-                TeamViewModel teamViewModel = new TeamViewModel();
-                teamViewModel.UserId = UserId;
-                teamViewModel.TeamId = TeamId;
-                teamViewModel.TeamRoleId = TeamRoleId;
+                TeamViewModel teamViewModel = new TeamViewModel
+                {
+                    UserId = UserId,
+                    TeamId = TeamId,
+                    TeamRoleId = TeamRoleId
+                };
                 TeamGroupRepository.DeleteTeamGroup(teamViewModel);
 
                 return RedirectToAction("Details", new { id = teamViewModel.TeamId });
